Crud user

+++ src/users/services/users.service.ts
@@ -1,10 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { DeepPartial, Repository } from 'typeorm';
 
 import { User } from './../entities/user.entity';
-
-import { CreateUserDto } from '../dtos/CreateUserDto';
+import { CreateOrUpdateUserDto } from '../dtos/CreateOrUpdateUserDto';
 
 @Injectable()
 export class UsersService {
@@ -16,25 +20,47 @@ export class UsersService {
     return this.userRepository.find();
   }
 
-  findOne(id: number) {
-    return this.userRepository.find({ where: { id: id } });
-  }
-
-  create(body: CreateUserDto) {
-    const newUser = this.userRepository.create(body);
-
-    return this.userRepository.save(newUser);
-  }
-
-  async update(id: number, body: any) {
+  async findOne(id: number) {
     const [user] = await this.userRepository.find({ where: { id: id } });
+    if (!user) throw new NotFoundException('Usuario no encontrado');
 
-    return this.userRepository.merge(user, body);
+    return user;
+  }
+
+  async create(body: CreateOrUpdateUserDto) {
+    try {
+      const user = this.userRepository.create(body as DeepPartial<User>);
+
+      return await this.userRepository.save(user);
+    } catch (err) {
+      if (err.code === '23505') {
+        throw new ConflictException('El email ya está en uso');
+      }
+      throw new InternalServerErrorException('Error al crear el usuario');
+    }
+  }
+
+  async update(id: number, body: CreateOrUpdateUserDto) {
+    const [user] = await this.userRepository.find({ where: { id: id } });
+    if (!user) throw new NotFoundException('Usuario no encontrado');
+
+    const merge = this.userRepository.merge(user, body as DeepPartial<User>);
+
+    return await this.userRepository.save(merge);
   }
 
   async delete(id: number) {
-    await this.userRepository.delete(id);
+    const [user] = await this.userRepository.find({ where: { id: id } });
+    if (!user) throw new NotFoundException('Usuario no encontrado');
 
-    return true;
+    try {
+      await this.userRepository.delete(id);
+
+      return { message: 'Usuario eliminado' };
+    } catch (err) {
+      throw new InternalServerErrorException(
+        `Error al eliminar usuario: ${err}`,
+      );
+    }
   }
 }
